feat(ui): fix image comparison slider resizing/aspect ratio jank
